pyproject.toml
[project]
name = "crewai_project"
version = "0.1.0"
description = "crewai_project using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai==0.159.0",
    "python-dotenv",
    "requests"
]

[project.scripts]
crewai_project = "crewai_project.main:run"
run_crew = "crewai_project.main:run"
train = "crewai_project.main:train"
replay = "crewai_project.main:replay"
test = "crewai_project.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
